+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## ChangeLog
 
+### 1.0.1
+
+* Added "throws error"
+* Added options default
+* Fix issue with init()
+
 ### 1.0.0
 
 * Initial release